pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audiogram-demo-clients"
version = "0.1.0"
description = "Demonstration clients for Audiogram gRPC API"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "grpcio==1.59.3",
    "protobuf==4.25.3",
    "python-keycloak==3.7.0",
    "click==8.1.7",
    "tabulate==0.9.0",
    "dynaconf[ini]==3.2.4",
    "python-dotenv==1.0.0",
    "ffmpeg-python>=0.2.0"
]

[project.scripts]
audiogram = "audiogram_cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["audiogram_client", "audiogram_cli"]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.7",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-protobuf>=4.24.0.20240311",
  "pytest>=8.0",
  "sphinx>=7.0",
  "sphinx-click>=5.1",
  "grpcio-tools>=1.59.3",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
]
ignore = [
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_optional = true
plugins = []
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]